requirements.txt
openai==1.3.0
pandas==2.1.0
python-dotenv==1.0.0
requests==2.31.0
tk==8.6.13
openpyxl==3.1.2
